support newer versions of numpy (fixes: #45)

+++ specvqgan/data/vggsound.py
@@ -45,7 +45,7 @@ class CropCoords(Crop):
 #         else:
 #             assert feat_len >= self.feat_sample_size
 #             # evenly spaced points
-#             idx = np.linspace(0, feat_len, self.feat_sample_size, dtype=np.long, endpoint=False)
+#             idx = np.linspace(0, feat_len, self.feat_sample_size, dtype=np.int32, endpoint=False)
 #             # xoooxooo -> ooxooxoo
 #             shift = feat_len // (self.feat_sample_size + 1)
 #             idx = idx + shift
@@ -63,7 +63,7 @@ class ResampleFrames(object):
         ## resample
         assert feat_len >= self.feat_sample_size
         # evenly spaced points (abcdefghkl -> aoooofoooo)
-        idx = np.linspace(0, feat_len, self.feat_sample_size, dtype=np.long, endpoint=False)
+        idx = np.linspace(0, feat_len, self.feat_sample_size, dtype=np.int32, endpoint=False)
         # xoooo xoooo -> ooxoo ooxoo
         shift = feat_len // (self.feat_sample_size + 1)
         idx = idx + shift
